Actualizada la lógica de formularios y vistas para mejorar la gestión de listas y perfiles de usuario. Se modificaron textos en las plantillas para una mejor experiencia de usuario y se implementó la lógica de restablecimiento de contraseña con un nuevo formulario.

+++ DJANGO-TAREAS/auth_app/models.py
@@ -41,6 +41,6 @@ def delete_old_pic(sender, instance, **kwargs):
         except Profile.DoesNotExist:
             return
         else:
-            if old_pic and old_pic != instance.pic:
+            if old_pic and old_pic != instance.pic and old_pic.name != 'profile_pics/default.jpg':
                 if os.path.isfile(old_pic.path):
                     os.remove(old_pic.path)
